refactor: add index files for merch constants and functions
