+++ apps/web/features/merch/constants/index.tsx
@@ -1,0 +1,2 @@
+export * from "./queryKeys"
+export * from "./routes"
