+++ apps/web/features/merch/functions/index.tsx
@@ -1,0 +1,3 @@
+export * from "./cart"
+export * from "./currency"
+export * from "./stock"
